requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
markdown>=3.5
